Move scroll hook to hooks file
